requirements.txt
# Core Django & REST Framework
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework-simplejwt==5.4.0

# PostgreSQL
psycopg2-binary==2.9.10

# Celery & Redis
celery==5.4.0
redis==5.2.0

# Image Processing
Pillow==10.4.0
pyvips==2.2.3

# Authentication & Security
pyotp==2.9.0
qrcode==7.4.2
argon2-cffi==23.1.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3
python-magic==0.4.27

# API Documentation
drf-yasg==1.21.7

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
